move Viewer view to frontend app
